Refactor project overwrite logic and enhance alert handling
Replaced inline overwrite logic with reusable functions for better modularity and maintainability. Improved alert dialog functionality to support third button options and streamlined callback handling in critical workflows.
