+++ src/js/actions/AlertModalActions.js
@@ -31,6 +31,7 @@ export function openAlertDialog(alertMessage, loading, buttonText = null, callba
  * @param {String} buttonLinkText - button text to show on left link button. (optional - if not present button is not added)
  * @param {function} callback2 - optional callback function that is called user clicks on button2Text.
  * @param {Boolean} notCloseableAlert - boolean to make the alert not closeable. (optional - if not present alert is closeable by clicking the x otherwise the x doesn't show up)
+ * @param {String} button3Text - button text to show for third button. (optional - if not present button is not added)
  * @return {Object} action content.
  */
 export function openOptionDialog(alertMessage, callback, button1Text, button2Text, buttonLinkText = null, callback2 = null, notCloseableAlert = false, button3Text = null) {
@@ -45,7 +46,7 @@ export function openOptionDialog(alertMessage, callback, button1Text, button2Tex
     notCloseableAlert,
     button3Text,
   };
-  console.log('alertConfig', alertConfig);
+  // console.log('alertConfig', alertConfig);
   return alertConfig;
 }
