+++ src/js/components/dialogComponents/Alert.js
@@ -81,7 +81,6 @@ class Alert extends Component {
       );
     }
 
-
     if (button1 && buttonLink) {
       buttonActions.unshift(
         <button
