+++ src/js/reducers/alertModalReducer.js
@@ -13,6 +13,14 @@ const initialState = {
   button3: null,
 };
 
+/**
+ * Redux reducer for managing alert modal state.
+ * Handles displaying alert dialogs with various button configurations and callbacks.
+ *
+ * @param {Object} state - Current state (defaults to initialState)
+ * @param {Object} action - Action object with type and payload
+ * @returns {Object} Updated state
+ */
 const alertModalReducer = (state = initialState, action) => {
   switch (action.type) {
   case consts.OPEN_ALERT_DIALOG:
@@ -50,8 +58,9 @@ const alertModalReducer = (state = initialState, action) => {
 export default alertModalReducer;
 
 /**
- * Checks if the alert dialog is open
- * @param state
- * @return {boolean}
+ * Checks if the alert dialog is currently closed (state matches initial state).
+ *
+ * @param {Object} state - The alert modal reducer state
+ * @returns {boolean} True if the state matches initialState (dialog is closed), false otherwise
  */
 export const getAlertIsOpen = state => state === initialState;
